autolm now working
- autolm now working. still need to add evaluation and perplexity measure
- turns out that torch transformers expect the sequence at dimension zero instead of batch unless you specify otherwise.
- added a bunch of default values for the config reader
- fixed progress bar updat values
- fixed some configs
- added tokenizers
