+++ .gitignore
@@ -4,6 +4,7 @@ debug/
 data/wackypedia_lemma/
 data/intermediate/
 checkpoints/
+.vscode
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
