+++ cfg/cv/mnist/mnist.yaml
@@ -22,6 +22,6 @@ data :
     cache_dir : "{{home}}/.cache/huggingface"
     shuffle : True
     batch_size : 2048
-    num_epochs : 10
+    num_epochs : 1
     eval_freq : 50
     log_freq : 10
