+++ mltoolkit/__main__.py
@@ -1,0 +1,2 @@
+from . import main
+main()
