+++ mltoolkit/trainers/__init__.py
@@ -1,5 +1,5 @@
 from .select import select
 from .example import TrainerExample
 from .base import TrainerBase
-from .mnist import TrainerMNIST
-from .autolm import TrainerAutoLM
+from .cv.mnist import TrainerMNIST
+from .nlp.autolm import TrainerAutoLM
