+++ mltoolkit/trainers/cv/__init__.py
@@ -1,0 +1,1 @@
+from .mnist import TrainerMNIST
