+++ mltoolkit/utils/__init__.py
@@ -5,5 +5,6 @@ from . import (
 	display,
     menus,
     data,
+    tokenizers,
 )
 from .chat_session import ChatSession
